fix: adds float cast to value
When trying to export the rows of resources which contain an advanced number field triggered an error "A non well formed numeric value encountered". This float cast fixes the issue.

This error can occure because of the maatwebsite/laravel-nova-excel package or just the way I've saved the value of the field within the database.

I'm still figuring it out but this was fast an easy fix.